Rename movies_details fragment
