+++ app/src/main/res/layout-land/fragment_movie_details.xml
