+++ app/src/main/res/layout/fragment_movie_details.xml
@@ -1,13 +1,14 @@
 <?xml version="1.0" encoding="utf-8"?>
-<ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.coordinatorlayout.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fillViewport="true">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
-        android:layout_height="wrap_content">
+        android:layout_height="match_parent">
 
         <ImageView
             android:layout_width="match_parent"
@@ -164,85 +165,20 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             app:barrierDirection="top"
-            app:constraint_referenced_ids="downey_ava,evans_ava,ruffalo_ava,hemsworth_ava"
+            app:constraint_referenced_ids="actors"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/cast" />
 
-        <ImageView
-            android:id="@+id/downey_ava"
-            style="@style/ActorAva"
-            android:layout_marginEnd="@dimen/activity_medium_margin"
-            android:src="@drawable/robert_downey_jr"
-            app:layout_constraintBottom_toTopOf="@id/downey_name"
-            app:layout_constraintEnd_toStartOf="@id/evans_ava"
-            app:layout_constraintStart_toEndOf="@id/guideline_left" />
-
-        <TextView
-            android:id="@+id/downey_name"
-            style="@style/ActorName"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:maxLines="2"
-            android:text="@string/robert_downey_jr"
-            app:layout_constraintEnd_toEndOf="@id/downey_ava"
-            app:layout_constraintStart_toStartOf="@id/downey_ava"
-            app:layout_constraintTop_toBottomOf="@id/downey_ava" />
-
-        <ImageView
-            android:id="@+id/evans_ava"
-            style="@style/ActorAva"
-            android:layout_marginEnd="@dimen/activity_medium_margin"
-            android:src="@drawable/chris_evans"
-            app:layout_constraintEnd_toStartOf="@id/ruffalo_ava"
-            app:layout_constraintStart_toEndOf="@id/downey_ava"
-            app:layout_constraintTop_toBottomOf="@id/barrier" />
-
-        <TextView
-            style="@style/ActorName"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:maxLines="2"
-            android:text="@string/chris_evans"
-            app:layout_constraintEnd_toEndOf="@id/evans_ava"
-            app:layout_constraintStart_toStartOf="@id/evans_ava"
-            app:layout_constraintTop_toBottomOf="@id/evans_ava" />
-
-        <ImageView
-            android:id="@+id/ruffalo_ava"
-            style="@style/ActorAva"
-            android:layout_marginEnd="@dimen/activity_medium_margin"
-            android:src="@drawable/mark_ruffalo"
-            app:layout_constraintEnd_toStartOf="@id/hemsworth_ava"
-            app:layout_constraintStart_toEndOf="@id/evans_ava"
-            app:layout_constraintTop_toBottomOf="@id/barrier" />
-
-        <TextView
-            style="@style/ActorName"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:maxLines="2"
-            android:text="@string/mark_ruffalo"
-            app:layout_constraintEnd_toEndOf="@id/ruffalo_ava"
-            app:layout_constraintStart_toStartOf="@id/ruffalo_ava"
-            app:layout_constraintTop_toBottomOf="@id/ruffalo_ava" />
-
-        <ImageView
-            android:id="@+id/hemsworth_ava"
-            style="@style/ActorAva"
-            android:src="@drawable/chris_hemsworth"
-            app:layout_constraintEnd_toStartOf="@id/guideline_right"
-            app:layout_constraintStart_toEndOf="@id/ruffalo_ava"
-            app:layout_constraintTop_toBottomOf="@id/barrier" />
-
-        <TextView
-            style="@style/ActorName"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:maxLines="2"
-            android:text="@string/chris_hemsworth"
-            app:layout_constraintEnd_toEndOf="@id/hemsworth_ava"
-            app:layout_constraintStart_toStartOf="@id/hemsworth_ava"
-            app:layout_constraintTop_toBottomOf="@id/hemsworth_ava" />
+        <androidx.recyclerview.widget.RecyclerView
+            android:id="@+id/actors"
+            android:layout_width="wrap_content"
+            android:layout_height="0dp"
+            android:orientation="horizontal"
+            tools:listitem="@layout/actor_item"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintTop_toBottomOf="@id/barrier"/>
 
     </androidx.constraintlayout.widget.ConstraintLayout>
-</ScrollView>
+</androidx.coordinatorlayout.widget.CoordinatorLayout>
